Move tsd tests to normal build
