+++ common/lib/core-interfaces/package.json
@@ -29,7 +29,6 @@
     "lint": "npm run eslint",
     "lint:fix": "npm run eslint:fix",
     "test": "npm run test:types",
-    "test:types": "tsd",
     "tsc": "tsc",
     "tsc:watch": "tsc --watch",
     "tsfmt": "tsfmt --verify",
@@ -55,12 +54,8 @@
     "eslint-plugin-react": "~7.22.0",
     "eslint-plugin-unicorn": "~26.0.1",
     "rimraf": "^2.6.2",
-    "tsd": "^0.17.0",
     "typescript": "~4.1.3",
     "typescript-formatter": "7.1.0"
-  },
-  "tsd": {
-    "directory": "test-d"
   },
   "typeValidation": {
     "version": "0.41.0",
