+++ packages/runtime/datastore-definitions/package.json
@@ -25,7 +25,6 @@
     "eslint:fix": "eslint --format stylish src --fix",
     "lint": "npm run eslint",
     "lint:fix": "npm run eslint:fix",
-    "test:types": "tsd",
     "tsc": "tsc",
     "tsfmt": "tsfmt --verify",
     "tsfmt:fix": "tsfmt --replace"
@@ -57,26 +56,8 @@
     "eslint-plugin-react": "~7.22.0",
     "eslint-plugin-unicorn": "~26.0.1",
     "rimraf": "^2.6.2",
-    "tsd": "^0.11.0",
     "typescript": "~4.1.3",
     "typescript-formatter": "7.1.0"
-  },
-  "tsd": {
-    "directory": "test-d",
-    "compilerOptions": {
-      "rootDir": "test-d",
-      "include": [
-        "test-d/*"
-      ],
-      "lib": [
-        "es2015"
-      ],
-      "types": [],
-      "exclude": [
-        "dist",
-        "node_modules"
-      ]
-    }
   },
   "typeValidation": {
     "version": "0.52.0",
